using test suite for file eav
