+++ cas_implementations/src/eav/file/mod.rs
@@ -157,7 +157,10 @@ pub mod tests {
 
     use eav::file::EavFileStorage;
     use holochain_core_types::{
-        cas::content::{AddressableContent, ExampleAddressableContent},
+        cas::{
+            content::{AddressableContent, ExampleAddressableContent},
+            storage::EAVTestSuite,
+        },
         eav::{EntityAttributeValue, EntityAttributeValueStorage},
     };
     use std::collections::HashSet;
@@ -170,171 +173,28 @@ pub mod tests {
         let entity_content = ExampleAddressableContent::from_content(&"foo".to_string());
         let attribute = "favourite-color".to_string();
         let value_content = ExampleAddressableContent::from_content(&"blue".to_string());
-        let eav = EntityAttributeValue::new(
-            &entity_content.address(),
-            &attribute,
-            &value_content.address(),
-        );
-        let mut eav_storage = EavFileStorage::new(temp_path).expect("should find holo file");;
-
-        assert_eq!(
-            HashSet::new(),
-            eav_storage
-                .fetch_eav(
-                    Some(entity_content.address()),
-                    Some(attribute.clone()),
-                    Some(value_content.address())
-                )
-                .expect("could not fetch eav"),
-        );
-
-        eav_storage.add_eav(&eav).expect("could not add eav");
-
-        let mut expected = HashSet::new();
-        expected.insert(eav.clone());
-        // some examples of constraints that should all return the eav
-        for (e, a, v) in vec![
-            // constrain all
-            (
-                Some(entity_content.address()),
-                Some(attribute.clone()),
-                Some(value_content.address()),
-            ),
-            // open entity
-            (None, Some(attribute.clone()), Some(value_content.address())),
-            // open attribute
-            (
-                Some(entity_content.address()),
-                None,
-                Some(value_content.address()),
-            ),
-            // open value
-            (
-                Some(entity_content.address()),
-                Some(attribute.clone()),
-                None,
-            ),
-            // open
-            (None, None, None),
-        ] {
-            println!("fetch");
-            assert_eq!(
-                expected,
-                eav_storage.fetch_eav(e, a, v).expect("could not fetch eav"),
-            );
-        }
+        EAVTestSuite::test_round_trip_test(
+            EavFileStorage::new(temp_path).unwrap(),
+            entity_content,
+            attribute,
+            value_content,
+        )
     }
 
     #[test]
     fn file_eav_one_to_many() {
         let temp = tempdir().expect("test was supposed to create temp dir");
         let temp_path = String::from(temp.path().to_str().expect("temp dir could not be string"));
-        let one = ExampleAddressableContent::from_content(&"foo".to_string());
-        // it can reference itself, why not?
-        let many_one = ExampleAddressableContent::from_content(&"foo".to_string());
-        let many_two = ExampleAddressableContent::from_content(&"bar".to_string());
-        let many_three = ExampleAddressableContent::from_content(&"baz".to_string());
-        let attribute = "one_to_many".to_string();
         let mut eav_storage = EavFileStorage::new(temp_path).unwrap();
-        let mut expected = HashSet::new();
-        for many in vec![many_one.clone(), many_two.clone(), many_three.clone()] {
-            let eav = EntityAttributeValue::new(&one.address(), &attribute, &many.address());
-            eav_storage.add_eav(&eav).expect("could not add eav");
-            expected.insert(eav);
-        }
-
-        // throw an extra thing referencing many to show fetch ignores it
-        let two = ExampleAddressableContent::from_content(&"foo".to_string());
-        for many in vec![many_one.clone(), many_three.clone()] {
-            eav_storage
-                .add_eav(&EntityAttributeValue::new(
-                    &two.address(),
-                    &attribute,
-                    &many.address(),
-                ))
-                .expect("could not add eav");
-        }
-
-        // show the many results for one
-        assert_eq!(
-            expected,
-            eav_storage
-                .fetch_eav(Some(one.address()), Some(attribute.clone()), None)
-                .expect("could not fetch eav"),
-        );
-
-        // show one for the many results
-        for many in vec![many_one.clone(), many_two.clone(), many_three.clone()] {
-            let mut expected_one = HashSet::new();
-            expected_one.insert(EntityAttributeValue::new(
-                &one.address(),
-                &attribute.clone(),
-                &many.address(),
-            ));
-            assert_eq!(
-                expected_one,
-                eav_storage
-                    .fetch_eav(None, Some(attribute.clone()), Some(many.address()))
-                    .expect("could not fetch eav"),
-            );
-        }
+        EAVTestSuite::test_one_to_many::<ExampleAddressableContent, EavFileStorage>(eav_storage)
     }
 
     #[test]
     fn file_eav_many_to_one() {
         let temp = tempdir().expect("test was supposed to create temp dir");
         let temp_path = String::from(temp.path().to_str().expect("temp dir could not be string"));
-        let one = ExampleAddressableContent::from_content(&"foo".to_string());
-        // it can reference itself, why not?
-        let many_one = ExampleAddressableContent::from_content(&"foo".to_string());
-        let many_two = ExampleAddressableContent::from_content(&"bar".to_string());
-        let many_three = ExampleAddressableContent::from_content(&"baz".to_string());
-        let attribute = "many_to_one".to_string();
-
         let mut eav_storage = EavFileStorage::new(temp_path).unwrap();
-        let mut expected = HashSet::new();
-        for many in vec![many_one.clone(), many_two.clone(), many_three.clone()] {
-            let eav = EntityAttributeValue::new(&many.address(), &attribute, &one.address());
-            eav_storage.add_eav(&eav).expect("could not add eav");
-            expected.insert(eav);
-        }
-
-        // throw an extra thing referenced by many to show fetch ignores it
-        let two = ExampleAddressableContent::from_content(&"foo".to_string());
-        for many in vec![many_one.clone(), many_three.clone()] {
-            eav_storage
-                .add_eav(&EntityAttributeValue::new(
-                    &many.address(),
-                    &attribute,
-                    &two.address(),
-                ))
-                .expect("could not add eav");
-        }
-
-        // show the many referencing one
-        assert_eq!(
-            expected,
-            eav_storage
-                .fetch_eav(None, Some(attribute.clone()), Some(one.address()))
-                .expect("could not fetch eav"),
-        );
-
-        // show one for the many results
-        for many in vec![many_one.clone(), many_two.clone(), many_three.clone()] {
-            let mut expected_one = HashSet::new();
-            expected_one.insert(EntityAttributeValue::new(
-                &many.address(),
-                &attribute.clone(),
-                &one.address(),
-            ));
-            println!("fetch");
-            assert_eq!(
-                expected_one,
-                eav_storage
-                    .fetch_eav(Some(many.address()), Some(attribute.clone()), None)
-                    .expect("could not fetch eav"),
-            );
-        }
+        EAVTestSuite::test_one_to_many::<ExampleAddressableContent, EavFileStorage>(eav_storage)
     }
 
 }
